+++ cache/147f3018_8440_4b05_82f1_b50d4bcac5e2/email_list.csv
@@ -1,0 +1,16 @@
+email,url
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/newsroom/events/zeiss-symposium.html
+do-not-reply@zeiss.com,https://www.zeiss.com/annualreport
+roundtable@deutsches-optisches-museum.de,https://www.deutsches-optisches-museum.de/en/
+Erinnerungen@D.O.M,https://www.deutsches-optisches-museum.de/en/
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/about-zeiss/responsibility/approach-to-responsibility.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/about-zeiss/history.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/innovation-and-technology/digital-innovation-partners.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/about-zeiss/history/zeiss-museum-of-optics.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/about-zeiss/responsibility.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/newsroom/press-photos.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/newsroom.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/about-zeiss/responsibility/products-and-value-chain.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/about-zeiss/responsibility/employees.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/about-zeiss/responsibility/social-engagement.html
+do-not-reply@zeiss.com,http://zeiss.com/corporate/int/about-zeiss/history/archives.html
