--- cache/7bf8b99a_7112_45ce_aecb_e73ebe18af13/email_list.csv
@@ -1,16 +1,0 @@
-email,url
-info@optovue.com,https://www.optovue.com/contact
-service@optovue.com,https://www.optovue.com/customer-care/support
-amy@amcpublicrelations.net,https://www.optovue.com/news/optovue-announces-expanded-fda-clearance-of-epithelial-thickness-mapping
-service@optovue.com,https://www.optovue.com/customer-care
-amy@amcpublicrelations.net,https://www.optovue.com/news/optovue-showcases-worlds-first-oct-angiography-metrics-and-non-contact-method-for-corneal-measurements-at-vision-expo-west
-christina_kirby@optovue.com,https://www.optovue.com/news-events
-amy@amcpublicrelations.net,https://www.optovue.com/news/optovue-highlights-worlds-first-octa-metrics-and-corneal-measurement-method-at-aao-2018
-academy@optovue.com,http://academy.optovue.com/
-privacy@sapho.com,https://www.optovue.com/legal/privacy-policy
-content@optovue.com,https://www.optovue.com/legal/privacy-policy
-service@optovue.com,https://www.optovue.com/customer-care/#usfoot
-info@optovue.com,https://www.optovue.com/international/contact
-service@optovue.com,https://www.optovue.com/customer-care/
-academy@optovue.com,http://academy.optovue.com
-amy@amcpublicrelations.net,https://www.optovue.com/international/news/optovue-announces-expanded-fda-clearance-of-epithelial-thickness-mapping
