+++ config.py
@@ -5,3 +5,5 @@ BLACKLIST = {
                 "URLS": ["whatsapp", "github", "bing", "facebook", "twitter", "flicker", "youtu", "google", "flickr", "commailto", "pinterest", "linkedin", "zencart", "wufoo", "youcanbook", "instagram"],
             }
 QUEUE = 100
+TIMEOUT_CONSTANT = 30e3
+VERSION = 0.3
